# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "消防巡检管理系统 API",
    "description": "基于 Node.js + Express + MongoDB 的消防设备巡检、报警处理、安全管理系统 API 文档。\n\n## 认证方式\n所有私有接口需在请求头中携带 Authorization: Bearer <token>。\n\n## 数据范围\n部分接口受数据范围权限控制：all（全部）/ department（本部门）/ self（仅本人）/ none（无权限）。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发环境"
    },
    {
      "url": "/",
      "description": "当前服务器"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "获取用户列表（分页）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "inactive", "locked"]
            }
          },
          {
            "name": "department",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "创建用户",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password"],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "realName": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "department": {
                    "type": "string"
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          },
          "400": {
            "description": "参数错误"
          }
        }
      }
    },
...
    "/api/roles": {
      "get": {
        "tags": ["Roles"],
        "summary": "获取角色列表（分页）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": ["Roles"],
        "summary": "创建角色",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "code"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^[A-Z_]+$"
                  },
                  "description": {
                    "type": "string"
                  },
                  "level": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          }
        }
      }
    },
...
    "/api/permissions": {
      "get": {
        "tags": ["Permissions"],
        "summary": "获取权限列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": ["Permissions"],
        "summary": "创建权限",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "code", "type", "module"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^(\\*|[a-z]+):(\\*|[a-z_]+)$"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["menu", "button", "api", "data"]
                  },
                  "module": {
                    "type": "string"
                  },
                  "parent": {
                    "type": "string"
                  },
                  "path": {
                    "type": "string"
                  },
                  "method": {
                    "type": "string",
                    "enum": ["GET", "POST", "PUT", "DELETE", "PATCH", "*"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          }
        }
      }
    },
...
    "/api/devices": {
      "get": {
        "tags": ["Devices"],
        "summary": "获取设备列表（分页）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "name": "deviceType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "fire_alarm",
                "sprinkler",
                "hydrant",
                "extinguisher",
                "smoke_detector",
                "heat_detector",
                "emergency_light",
                "evacuation_sign",
                "fire_door",
                "other"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["normal", "warning", "fault", "offline", "maintenance", "scrapped"]
            }
          },
          {
            "name": "building",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "floor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": ["Devices"],
        "summary": "创建设备",
        "description": "设备编号自动生成，格式为 类型前缀-年份-序号（如 FA-2026-0001）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["deviceName", "deviceType"],
                "properties": {
                  "deviceName": {
                    "type": "string"
                  },
                  "deviceType": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "manufacturer": {
                    "type": "string"
                  },
                  "location": {
                    "type": "object"
                  },
                  "installDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "expiryDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "checkCycle": {
                    "type": "integer",
                    "description": "检查周期（天）"
                  },
                  "remark": {
                    "type": "string"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          }
        }
      }
    },
...
    "/api/alarms": {
      "get": {
        "tags": ["Alarms"],
        "summary": "获取报警列表（分页）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "processing", "resolved", "false_alarm", "cancelled"]
            }
          },
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["info", "warning", "critical", "emergency"]
            }
          },
          {
            "name": "alarmType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "smoke",
                "temp_abnormal",
                "manual_button",
                "phone_report",
                "patrol_find",
                "other"
              ]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/api/inspections": {
      "get": {
        "tags": ["Inspections"],
        "summary": "获取巡检列表（分页）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "in_progress", "completed", "overdue", "cancelled"]
            }
          },
          {
            "name": "inspectionType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["daily", "weekly", "monthly", "quarterly", "annual", "special"]
            }
          },
          {
            "name": "assignedTo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": ["Inspections"],
        "summary": "创建巡检计划",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "inspectionType",
                  "planStartTime",
                  "planEndTime",
                  "checkItems"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "inspectionType": {
                    "type": "string",
                    "enum": ["daily", "weekly", "monthly", "quarterly", "annual", "special"]
                  },
                  "planStartTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "planEndTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "assignedTo": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "mongodb-id"
                    }
                  },
                  "devices": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "locations": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "checkItems": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    },
                    "minItems": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "remark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          }
        }
      }
    },
...
    "/api/reports/devices": {
      "get": {
        "tags": ["Reports"],
        "summary": "获取设备报表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/reports/alarms": {
      "get": {
        "tags": ["Reports"],
        "summary": "获取报警报表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/reports/inspections": {
      "get": {
        "tags": ["Reports"],
        "summary": "获取巡检报表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回按状态、结果、执行人统计"
          }
        }
      }
    },
...
  }
}